+++ py/prospect/scripts/prospect_pages.py
@@ -162,9 +162,12 @@ def load_spectra_zcat_from_dbentry(db_entry, args, log, with_redrock_version=Tru
             if os.path.isfile(os.path.join(the_dir, "redrock-"+file_label+".fits")):
                 redrock_is_pre_everest = False
                 the_zcat = Table.read(os.path.join(the_dir, "redrock-"+file_label+".fits"), 'REDSHIFTS')
-            else: # pre-everest Redrock file nomenclature
+            else:  # pre-everest Redrock file nomenclature
                 redrock_is_pre_everest = True
                 the_zcat = Table.read(os.path.join(the_dir, "zbest-"+file_label+".fits"), 'ZBEST')
+            if hasattr(the_zcat['SUBTYPE'], 'mask'):  # work around Table auto-masking in astropy 5
+                blanksubtype = the_zcat['SUBTYPE'].mask
+                the_zcat['SUBTYPE'][blanksubtype] = ''
             if with_redrock_version:
                 if redrock_is_pre_everest:
                     hdulist = astropy.io.fits.open(os.path.join(the_dir, "zbest-"+file_label+".fits"))
@@ -310,9 +313,12 @@ def main():
             spectra = desispec.io.read_spectra(args.spectra_files[i_file])
             if args.zcat_files is not None:
                 try:
-                    zcat = Table.read(args.zcat_files[i_file],'REDSHIFTS')
-                except KeyError: # pre-everest Redrock file nomenclature
-                    zcat = Table.read(args.zcat_files[i_file],'ZBEST')
+                    zcat = Table.read(args.zcat_files[i_file], 'REDSHIFTS')
+                except KeyError as e:  # pre-everest Redrock file nomenclature
+                    zcat = Table.read(args.zcat_files[i_file], 'ZBEST')
+                if hasattr(zcat['SUBTYPE'], 'mask'):  # work around Table auto-masking in astropy 5
+                    blanksubtype = zcat['SUBTYPE'].mask
+                    zcat['SUBTYPE'][blanksubtype] = ''
                 #- Add redrock version to zcat
                 hdulist = astropy.io.fits.open(args.zcat_files[i_file])
                 zcat['RRVER'] = hdulist[hdulist.index_of('PRIMARY')].header['RRVER']